+++ flask_unchained/bundles/sqlalchemy/services/session_manager.py
@@ -1,3 +1,4 @@
+from contextlib import contextmanager
 from flask_unchained import BaseService, injectable
 from typing import *
 
@@ -29,6 +30,16 @@ class SessionManager(BaseService):
 
     def commit(self):
         self.db.session.commit()
+
+    @property
+    @contextmanager
+    def no_autoflush(self):
+        autoflush = self.db.session.autoflush
+        self.db.session.autoflush = False
+        try:
+            yield self
+        finally:
+            self.db.session.autoflush = autoflush
 
     def __getattr__(self, method_name):
         return getattr(self.db.session, method_name)