+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 * make the `ModelManager.model` class attribute a meta option
 * rename the `flask db drop --drop` option to `flask db drop --force` to skip prompting
 * rename the `flask db reset --reset` option to `flask db reset --force` to skip prompting
+* add `no_autoflush` to `SessionManager`
 
 ## 0.5.1 (2018/07/25)